+++ hyperdbg/script-eval/header/ScriptEngineCommonDefinitions.h
@@ -299,7 +299,7 @@ static const char *const RegistersNames[] = {
 #define PSEUDO_REGISTER_IP 8
 #define PSEUDO_REGISTER_BUFFER 9
 #define PSEUDO_REGISTER_CONTEXT 10
-#define PSEUDO_REGISTER_EVENT_TAG 11
-#define PSEUDO_REGISTER_EVENT_ID 12
+#define PSEUDO_REGISTER_TAG 11
+#define PSEUDO_REGISTER_ID 12
 
 #endif